backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
Django==5.2
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
django-jazzmin==3.0.1
django-modeltranslation==0.19.14
django-rest-swagger==2.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
netaddr==1.3.0
openapi-codec==1.3.2
packaging==25.0
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.0
python-openid2 @ git+https://github.com/openid/python-openid.git@afa6adacbe1a41d8f614c8bce2264dfbe9e76489
python3-openid @ git+https://github.com/necaris/python3-openid.git@fbc7f6a2dc5159b890119a044215a6aee4cefc8d
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
wrapt==1.17.3
yookassa==3.6.0